Remove status pass from server, and test this.
This commit removes the status check pass from the demo server, as
that pass was made too expensive and time-consuming with the new,
eventually consistent system.  The submit pass does all the work
now.

This commit also includes a copy of the "house_example.json" file
from the Bowtie SDK Documentation, and includes documentation on
how to send that file to the demo server to test that the server
is communicating with the Young Alfred sandbox correctly.