Replace submit/status cycle with a single submit call.
This commit replaces the old submit/wait for status cycle with the
new shallow-validation, one-submit-only process currently supported
by the public API.  Fixing this required changes to both the client
and the server, but those changes resulted in a much more streamlined
demo that I think will be easier for partners to understand.

Since repeated testing of the cycle in the whole was a pain in the, er,
next, this commit also provides a complete, functional end-to-end
test of the Young Alfred application process, showing how the system
behaves under ideal conditions.  The application provided is the
all-options-on variant from the Bowtie SDK documentation, and the
WebDriverIO test rips through the entire application in about 20
seconds.

Lessons learned: the odd focus issues caused by the constant refresh
of the HTML document leads to some interesting testing requirements,
the most fundamental of which is the re-acquisition of the body root
after every iteration, and the need to focus (via `click`) a text
input before using it, and to trigger an updated to the SDK by losing
focus (via `tab`) after the text field is filled out.  These events
usually happen as the user mouses or scrolls so this is not a problem
in the field, but the test is doing neither, so these extras simulate
that behavior.